Add OracleInterface and CentralizedOracle implementation (#87)
The implementation allows the owner to push prices for (symbol, time)
pairs that have been requested.

The contract holds on to the requests, and verifies that you can't
push prices that haven't been requested. Events are also fired whenever a price is requested and when a price is available.

Fixes #85 and addresses #86.