Change `hasRevealed` to only apply to the current round (#555)
For example, if a price request fails to resolve, then it automatically
becomes pending in the next round. Then, suppose the next commit phase
goes by with no activity. The voter dapp will then try to pull
the `hasRevealed` status, which should show `Cannot be revealed`. The
voter dapp relies on the `hasRevealed` call to return `false`.