Token preapproval (#671)
This is pretty standard outside of the modification of the scss.

I had to modify it because the span and strong components were inheriting properties from the context. Specifically, the span was inheriting a red color, making the text invisible on top of the button. Using divs, the text styling may be slightly different than the other processing buttons, but everything is visible, in the expected color, and works.

Feel free to tweak the above as you see fit as it's certainly not a perfect solution.

Note: I won't be able to respond to feedback on this PR before they need to be submitted, so if you do have requested changes, feel free to just copy my branch: https://github.com/mrice32/protocol/tree/token_preapproval and modify it to create new PR with the fixes.

Fixes #650.