Refactor Voting supportedIdentifier struct and methods (#819)
Created new IdentifierWhitelist contract to manage storage of the price identifiers that the Oracle supports. OracleInterface no longer has isIdentiferSupported method and Oracle implementations (Voting, MockOracle) no longer have addSupportedIdentifier and removeSupportedIdentifier.

Added an Interface for IdentifierWhitelist so that implementations of OracleInterface don't have to import the entire bytecode of the IdentifierWhitelist implementation. 

Governor contract now needs to know the address for the IdentifierWhitelist so that it can add/remove identifiers. 

Deployment scripts of Voting now has a prerequisite to deploy an IdentifierWhitelist and pass it as reference upon construction of Voting. 

IdentiferWhitelist must be added to Finder. 

Add IdentifierWhitelist to list of addresses available in Finder, primarily so that Governor is aware of Identifier and can add identifiers.

JS tests and scripts now must call IdentifierWhitelist (previously Voting/MockOracle) to add/removeSupportedIdentifiers. Added unit tests for new IdentifierWhitelistContract

Updated docs to reflect change of OracleInterface.

Added address of IdentifierWhitelist on testnet to Registry

Changed sponsor-dapp-v2 to call IdentifierWhitelist to check if an identifier is supported.

Added IdentifierWhitelist to drizzle config.

Signed-off-by: Nick Pai <npai.nyc@gmail.com>